organize APIs code structure
