+++ aspic/core/config.py
@@ -1,0 +1,7 @@
+from pydantic import BaseSettings
+
+class Settings(BaseSettings):
+    API_V1_STR: str = "/api/v1"
+    PROJECT_NAME: str = "ASPIC OPERATOR"
+
+settings = Settings()
