+++ pyproject.toml
@@ -11,6 +11,7 @@ click = "^8.1.2"
 PyYAML = "^6.0"
 pykube = "^0.15.0"
 fastapi = "^0.75.1"
+pydantic = "^1.9"
 uvicorn = {extras = ["standard"], version = "^0.17.6"}
 
 [tool.poetry.dev-dependencies]
